Templating

+++ build/build.sh
@@ -13,10 +13,14 @@ manubot \
   --cache-directory=ci/cache \
   --log-level=DEBUG
 
+BUILD_HTML="false"
+BUILD_PDF="false"
+BUILD_DOCX="true"
+
 # pandoc settings
 # Exports so that we can convert and resize figures.
-CSL_PATH=build/assets/style.csl
-DOCX_PATH=build/assets/pandoc-reference.docx
+CSL_PATH=build/assets/journal-of-chemical-theory-and-computation.csl
+DOCX_PATH=build/assets/pandoc.docx
 SVG_FIX=build/assets/pandoc-svg.py
 export PNG_RESIZE=build/assets/png_resize.py
 BIBLIOGRAPHY_PATH=output/references.json
@@ -27,48 +31,53 @@ mkdir -p output
 
 # Create HTML output
 # http://pandoc.org/MANUAL.html
-echo "Exporting HTML manuscript"
-pandoc --verbose \
-  --from=markdown \
-  --to=html5 \
-  --filter=pandoc-fignos \
-  --filter=pandoc-eqnos \
-  --filter=pandoc-tablenos \
-  --bibliography=$BIBLIOGRAPHY_PATH \
-  --csl=$CSL_PATH \
-  --metadata link-citations=true \
-  --mathjax \
-  --css=github-pandoc.css \
-  --include-in-header=build/assets/analytics.html \
-  --include-after-body=build/assets/anchors.html \
-  --include-after-body=build/assets/hypothesis.html \
-  --output=output/manuscript.html \
-  $INPUT_PATH
+if [ "$BUILD_HTML" = "true" ];
+then
+  echo "Exporting HTML manuscript"
+  pandoc --verbose \
+    --from=markdown \
+    --to=html5 \
+    --filter=pandoc-fignos \
+    --filter=pandoc-eqnos \
+    --filter=pandoc-tablenos \
+    --bibliography=$BIBLIOGRAPHY_PATH \
+    --csl=$CSL_PATH \
+    --metadata link-citations=true \
+    --mathjax \
+    --css=github-pandoc.css \
+    --include-in-header=build/assets/analytics.html \
+    --include-after-body=build/assets/anchors.html \
+    --include-after-body=build/assets/hypothesis.html \
+    --output=output/manuscript.html \
+    $INPUT_PATH
+fi
 
 # Create PDF output
-echo "Exporting PDF manuscript"
-ln -s content/images images
-pandoc \
-  --from=markdown \
-  --to=html5 \
-  --pdf-engine=weasyprint \
-  --pdf-engine-opt=--presentational-hints \
-  --filter=pandoc-fignos \
-  --filter=pandoc-eqnos \
-  --filter=pandoc-tablenos \
-  --bibliography=$BIBLIOGRAPHY_PATH \
-  --csl=$CSL_PATH \
-  --metadata link-citations=true \
-  --webtex=https://latex.codecogs.com/svg.latex? \
-  --css=webpage/github-pandoc.css \
-  --output=output/manuscript.pdf \
-  $INPUT_PATH
-rm -r images
+if [ "$BUILD_PDF" = "true" ];
+then
+  echo "Exporting PDF manuscript"
+  ln -s content/images images
+  pandoc \
+    --from=markdown \
+    --to=html5 \
+    --pdf-engine=weasyprint \
+    --pdf-engine-opt=--presentational-hints \
+    --filter=pandoc-fignos \
+    --filter=pandoc-eqnos \
+    --filter=pandoc-tablenos \
+    --bibliography=$BIBLIOGRAPHY_PATH \
+    --csl=$CSL_PATH \
+    --metadata link-citations=true \
+    --webtex=https://latex.codecogs.com/svg.latex? \
+    --css=webpage/github-pandoc.css \
+    --output=output/manuscript.pdf \
+    $INPUT_PATH
+  rm -r images
+fi
 
 # Create DOCX output when user specifies to do so
 if [ "$BUILD_DOCX" = "true" ];
 then
-    echo "Exporting Word Docx manuscript"
     pandoc --verbose \
     --from=markdown \
     --to=docx \
@@ -79,11 +88,12 @@ then
     --filter=$SVG_FIX \
     --bibliography=$BIBLIOGRAPHY_PATH \
     --csl=$CSL_PATH \
-    --metadata link-citations=true \
     --reference-doc=$DOCX_PATH \
+    --metadata link-citations=true \
     --resource-path=.:content \
     --output=output/manuscript.docx \
     $INPUT_PATH
+
 fi
 
 echo "Build complete"
